backend/requirements.txt
django
djangorestframework
psycopg2-binary
celery
django-cors-headers
